Enable dark mode and adjust hljs and SVG styles
This adds support for OS-wide user preferences and enables the
ReSpec-provided user override of the same,

The adjusted styles also fixes the current problem where OS-wide dark
mode is applied on the syntax highlight blocks, even though the document
was locked to light mode.

The color-scheme property controls selection of color-scheme also in
external SVG documents, so the embedded SVG is replaced by an external
include through an object element.

This should address the problems reported in #22.